summary: remove psycopg2 usage
Psycopg2 data types do not generally
leak through SQLAlchemy. Since this
is an internal helper function, remove
the usage of psycopg2. Even if the psycopg2
data types are leaking through SQLAlchemy,
there is a single call to the function
from test_page_loads, and that function
can live with a less pretty range printout.